Move test utils to another file
