+++ corehq/form_processor/interfaces.py
@@ -1,12 +1,11 @@
 from couchdbkit import ResourceNotFound
 from casexml.apps.case.dbaccessors import get_reverse_indices_for_case_id
 from casexml.apps.case.util import get_case_xform_ids, post_case_blocks
-from casexml.apps.phone.models import SyncLog
 from corehq.apps.hqcase.dbaccessors import get_case_ids_in_domain
 from corehq.util.test_utils import unit_testing_only
 
 from dimagi.utils.couch.undo import DELETED_SUFFIX
-from dimagi.utils.couch.database import iter_docs, safe_delete
+from dimagi.utils.couch.database import iter_docs
 from casexml.apps.case.models import CommCareCase
 from couchforms.util import process_xform
 from couchforms.models import doc_types, XFormInstance, XFormError
@@ -185,30 +184,3 @@ class FormProcessorInterface(object):
         from casexml.apps.case.cleanup import safe_hard_delete
         case = cls._get_case(case_generic.id)
         safe_hard_delete(case)
-
-    @classmethod
-    @unit_testing_only
-    def delete_all_cases(cls):
-        cls._delete_all(CommCareCase.get_db(), 'case/get_lite')
-
-    @classmethod
-    @unit_testing_only
-    def delete_all_xforms(cls):
-        cls._delete_all(XFormInstance.get_db(), 'couchforms/all_submissions_by_domain')
-
-    @classmethod
-    @unit_testing_only
-    def delete_all_sync_logs(cls):
-        cls._delete_all(SyncLog.get_db(), 'phone/sync_logs_by_user')
-
-    @staticmethod
-    def _delete_all(db, viewname):
-        deleted = set()
-        for row in db.view(viewname, reduce=False):
-            doc_id = row['id']
-            if id not in deleted:
-                try:
-                    safe_delete(db, doc_id)
-                    deleted.add(doc_id)
-                except ResourceNotFound:
-                    pass
